Fixes - Openstack HA , move the external VIP along with internal VIP and vice-versa
In a HA setup, when external VIP and internal VIP are present, when one VIP moves
the other does not move. In the case, when the control/data interface goes down,
the internal VIP moves to the next node, but the managment VIP stays in the same
host as before. Hence, any request to that managment VIP does not work.

The solution is to move the VIPs together by creating a VRRP group.

Change-Id: I2d56515e7cf392dc0ac12e23edc4b9e35a8d2b50
ClosesBug:1381192