Populating SCHEDULER section in the service monitor config file
Currently the service monitor uses loopback address as the collector ip.
if collector node is different than the cfgm node, It fails to get the vrouter node status
for launcing network name space. Fixing it by populating SCHEDULER section in its config file
which is used by the service monitor.

Change-Id: Ia0fbc5f9fc64cfa45b30e639f484e956b8e40d2b