Multi-pool support with Ceph storage
Closes-Bug: #1378966
Feature: The 'Ganges' project required the support of having multiple
HDD pools so that the replica of an object will not be placed on the
same chassis, which has 4 nodes. Disks in each node will be placed in
a different pool.
Change: Added support for the above. Configuration option added in the
'disks' and the 'ssd-disks' option in the testbed.py to have additional
pool name. Ceph pools are created based on the pool name. The crush
map is modified by allocating OSDs to the corresponding pools based on
the testbed.py configuration.
Test: QA Sanity passed for existing configuraiton.
      Verified by creating 2 HDD and 2 SSD pools.

Change-Id: Ie52025269e799a6cedefbb6bc85cc6688d6aa67d