Fixes: Openstack HA , move the external VIP along with internal VIP and vice-vesa
Changed Value of max_overflow to -1. This is to allow max concurrent connections
There are multiple bug fixes for this commit.
Added new scheduler config to have the max attempts to create instance.
Keepalived config to support combined external VIP movement to the node where internal VIP moves to -
Closes-Bug: #1381192.

Change in WSREP config for innodb_log_file_size and update offset to 100+ -
Closes-Bug: #1385593

Change-Id: If30855f21079b3b803961694c730db12fe86a8ff