Fixes: Openstack HA , cmon monitoring script is not able to start mysql when mysql is down after few days
Token cleanup and CMON logs cleanup is scheduled to run as a
cronjob in the midnight in all the three OpenStack controllers.
If the db is huge, it leads to WSREP lock issues leading to
data inconsistency. WSREP kills MySQL when there is data inconsistency
which leads to MySQL being killed.

To fix this, we will space out the cleanup job to run every hour.

Change-Id: I3d0375b3c731aeb7dbccfea2d955fd286e5d3146
Closes-Bug:1408756