Fixes: Openstack HA , mysql gets locked up during node failure test.
We have a cleanup script which cleans up CMON and Keystone tokens
every night. CMON produces a lot of logs (recorded into the database)
and hence the CMON DB in MySQL is pretty huge. When the CMON purge
script runs it results in data inconsistency between MySQL nodes
which is known to happen for DELETE operations in large MySQL tables.

The following change prevents this by doing the following -
* The time interval which CMON logs data is increased from 30 seconds
to 6 hours.
* Running CMON purge operation only in a node where CMON is running.
* Reverting a previous change which spaced out running the CMON commands

Change-Id: Ia8fef70a0d60cd59e011ebd9ac9c6e9514f354bd
Closes: 1414245