Fixes: Openstack HA , mysql gets locked up during node failure test.
We have a cleanup script which cleans up CMON and Keystone tokens
every night. CMON produces a lot of logs (recorded into the database)
and hence the CMON DB in MySQL is pretty huge. When the CMON purge
script runs it results in data inconsistency between MySQL nodes
which is known to happen for DELETE operations in large MySQL tables.

The following change prevents this by doing the following -
* The time interval which CMON logs data is increased from 30 seconds
to 24 hours.
* Running CMON purge operation only in a node where CMON is running.
* Running cleanup every hour

Change-Id: I6d855b26c59ca6896a50fee60fff2f62097f9bf9
Closes: 1414245