1) cassandra_name changed to cassandra_user 2) config/common.py, daemons are made to take conf file as argument tomake it uniform 3) In collector, templates created for collector,query and api and populated if authentication is on 4) In database, OS version check to create user is removed
collector:

Read the arguments passed
create the /etc/contrail/contrail-database.conf if it does not exist

config:
Read the arguments and create the contrail-database.conf if it does not exist.
Adjust the supervsor ini file for device manager, contrail-api, discovery,  to take contrail-database.conf as an additional argument

when running on openstack controller, pass the contrail-database.conf as an additional
argument when invoking schema, api, device-manager, svc-monitor

Database:
If cassandra_name argument is passed, create the user account in cassandra using cqlsh.
Included appropriate retry.Also change the cassandra.yaml to turn on authentication

Partial-Bug: #1392115
Conflicts:

	contrail_provisioning/collector/setup.py

Change-Id: I296a09f953474236c0641cd19090cd8fd2e62e52