Fix fetching bond members to remove from config.
Bond members have to be taken both from --members and from Linux.
In case we want to remove one of members from an existing bond,
we need to remove all members from config file, not just the
ones passed in --members, which are the ones that are going to be
used. Old member (that was not passed in --members) has to be
removed as well. On the next stage, only new members will be appended
to config.

Change-Id: I0b36b604f210bd649fe2f6db725cf8fc917079a8
Closes-bug: #1547562