Fixes: Openstack HA : mysql is not started after node failure of openstack node
CMON monitor script is responsible for starting CMON process based on the Zookeeper
lock that is acquired by one of the three monitoring process in one of the three
nodes.

However, the process is started using "service cmon start" which internally uses
the /etc/init.d/cmon script which is read only.

Making the script executable will start the CMON process.

Change-Id: I294a8cfbc46ca529ff4eb11c357dcd15a9f3533c
ClosesBug: 1541304