Support configuring PCI Virtual Function ifaces
In order for the VFs to work correctly, their parent (the physical) interface
has to be up. We achieve this by allowing to put a no-ip interface entry in
/etc/network/interfaces.

Also, we have to set the VFs MAC address such that it matches the MAC of the
physical interface.

Moreover, we set explicitly the MAC address of the vhost0 interface to match
the address of the interface taken by the vRouter. This is required when using
the DPDK vRouter over the VF interface. Without this the vhost0 would get
random MAC that breaks vRouter operation.

Change-Id: Ib0db3cc73b3e448591ea5cca31605467969e713e
Partial-Bug: #1564935