Fix generation of pci_passthrough_whitelist nova.conf option
When the DPDK vRouter uses one of the PCI Virtual Functions, the particular
function can not be used for SRIOV ports at the same time.

This commit ensures that the PCI address taken by the DPDK vRouter will not be
included in pci_passthrough_whitelist options, which effectively fulfills the
requirement stated above.

Change-Id: Ia04ee8433768bb5e96d647e976a779c42ad8bbb4
Partial-Bug: #1564935