Partial-Bug: #1593008
contrail-database [aka cassandra] upgrade needs to be handled a bit
differently as the database needs migration using upgradesstables
when going from 1 release to another. For RPM, the package names
themselves get changed, hence that needs to be handled appropriately.
For any 2 pair of upgrades, we maintain a list of intermediate packages
that we need to successively go through and run upgradesstables.

Change-Id: Id1f34ca186383f79700c4c385380599c0aa5c58b