Merge below 3.0 changes into master.
- When provisioning a compute node with VFs, use the same command
as setup_interface.

- Avoid bringing down the PF when the VF comes up. Otherwise, addition
of the VF to a bond fails, probably due a bug in the ixgbe driver.

- Modify /etc/network/interfaces to add the right commands for VFs

- Enslave VFs explicitly in /etc/network/interfaces to work around flaky
behavior in ixgbe diver.

- Prevent prompt for user/password when running fab setup_all

- vlan interface creation fails if the pre-up does "ifconfig up" of the vlan
interface that is being brought up. This is not an issue for physical interfaces,
but fails if the VLAN is created on top of SR-IOV VFs. post-up should be
used instead of pre-up as the interface would have been created by then.

- Bringing up interfaces one at a time intstead of all.

Change-Id: I55a583e8e89a31c3b313ad223bae2976e3f344f3
Closes-Bug: 1573255 1585876