Provide the default filters in SRIOV config
While setting up the SRIOV configuration in nova.conf for nova
scheduler, the scheduler_default_filters is currently setup wth only
PciPassthroughFilter. This results in some default filters not being
applied which are initialised as part of nova scheduler intialisation.
One such filter is availability zone filter. If
scheduler_default_filters is initialised explictly from nova.conf it
should contain all the filter the nova-scheduler is initialising as part
of its code. This list of filters varies from one openstack release to
another.

The provisioning script is modified to provide this list explicitly. The
list is hard coded as per the openstack release

Change-Id: I88db0490237a92677ecf655132f980632f52e1e6
closes-bug: #1567000