Fixes: Openstack HA : cmon is not bringing up mysql after node failure tests
Starting CMON is controlled by a script which acquires a zookeeper locker
and runs only one instance of CMON. Hence CMON will be started by the script
which takes the lock. In another review, I have removed the startup script
which runs cmon on boot. Hence, removing the workaround to make the cmon startup
script as non-executable which was preventing cmon from running.

Change-Id: I39f8296fa8ae7e330b6e544edaf50fd227a336c6
Closes-Bug: 1634642