Fixes: Launching of VMs getting failed
In Kiloi and Liberty, the neutron auth_url should be mentined as admin_auth_url.
It is changed to auth_url in Mitaka and later releases. Because of this,
the auth_url is poitning to the default value localhost:5000.

Fix is to set the neutron auth_url based on the release.

Change-Id: Iabd2ba7716072a102c09eaee0ccff76200d87cfe
Closes-Bug: 1642005