Change section 'DEFAULT' to 'DEFAULTS' in contrail-database-nodemgr.conf
python ConfigParser treats section 'DEFAULT' as special and
whatever is in the 'DEFAULT' section is propagated to every other
section. Due to this, the current code in nodemgr is working only
as a result of parsing the 'DISCOVERY' section and it fails once
the 'DISCOVERY' section is removed.

Handle upgrade by replacing the contents in contrail-database-nodemgr.conf
under section 'DEFAULT' and moving to section 'DEFAULTS'.

Change-Id: I9f74699c13ae88ccbbd66e5cd35e4d59ce97a5fb
Closes-Bug: #1660535