Set novncproxy port appropriately for Ocata
The default value of novncport is 6080. Always use the default values. It
is not clear why the value was changed to 5999 in the contrail compute
setup. If this value is changed to a different value from 6080, then the
ocata configs should be overridden by setting the nova_vnc_port to the
same value in kolla_globals.yml (or kolla_globals in the cluster json)

Console for instances not accessible in Ocata

In Ocata (and above), the novncproxy server listens on the control-data IP and
not on the management IP. So the novncproxy_base_url should be pointing to the
control-data-ip and not the management IP.

Change-Id: Ie08aa456e569e317c38d898d6e79c5de9b7c467b
Closes-bug: #1725026
Closes-bug: #1722549