Set novncproxy port appropriately for Ocata
The default value of novncport is 6080. Always use the default values. It
is not clear why the value was changed to 5999 in the contrail compute
setup. If this value is changed to a different value from 6080, then the
ocata configs should be overridden by setting the nova_vnc_port to the
same value in kolla_globals.yml (or kolla_globals in the cluster json)

Change-Id: Ie08aa456e569e317c38d898d6e79c5de9b7c467b
Closes-bug: #1722549